protoc-plugin: Move all declarations into model.kt
