--- protoc-gen/src/main/kotlin/kotlinx/rpc/protobuf/model/FileDeclaration.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright 2023-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.rpc.protobuf.model
-
-data class FileDeclaration(
-    val name: String,
-    val packageName: FqName.Package,
-    val dependencies: List<FileDeclaration>,
-    val messageDeclarations: List<MessageDeclaration>,
-    val enumDeclarations: List<EnumDeclaration>,
-    val serviceDeclarations: List<ServiceDeclaration>,
-    val deprecated: Boolean,
-    val doc: String?,
-)
