--- protoc-gen/src/main/kotlin/kotlinx/rpc/protobuf/model/MessageDeclaration.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright 2023-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.rpc.protobuf.model
-
-data class MessageDeclaration(
-    val outerClassName: FqName,
-    val name: FqName,
-    val actualFields: List<FieldDeclaration>, // excludes oneOf fields, but includes oneOf itself
-    val oneOfDeclarations: List<OneOfDeclaration>,
-    val enumDeclarations: List<EnumDeclaration>,
-    val nestedDeclarations: List<MessageDeclaration>,
-    val deprecated: Boolean,
-    val doc: String?,
-)
