--- protoc-gen/src/main/kotlin/kotlinx/rpc/protobuf/model/MethodDeclaration.kt
@@ -1,13 +1,0 @@
-/*
- * Copyright 2023-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.rpc.protobuf.model
-
-data class MethodDeclaration(
-    val name: String,
-    val clientStreaming: Boolean,
-    val serverStreaming: Boolean,
-    val inputType: Lazy<MessageDeclaration>,
-    val outputType: Lazy<MessageDeclaration>,
-)
