--- protoc-gen/src/main/kotlin/kotlinx/rpc/protobuf/model/ServiceDeclaration.kt
@@ -1,10 +1,0 @@
-/*
- * Copyright 2023-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.rpc.protobuf.model
-
-data class ServiceDeclaration(
-    val name: FqName,
-    val methods: List<MethodDeclaration>,
-)
